feat(#320): rename client-test-utils rpc-msw to orpc, drop test-setup
The rpc-msw subpath is renamed orpc for a plainer name, and the
test-setup subpath is removed now that registerMSWLifecycle lives in
@vers/test-utils/bun. app-web and lib-email pick up the new dependency
for their own MSW-lifecycle preload wiring.
