+++ projects/app-web/package.json
@@ -76,6 +76,7 @@
     "@types/react-dom": "catalog:",
     "@types/three": "catalog:",
     "@vers/client-test-utils": "workspace:*",
+    "@vers/test-utils": "workspace:*",
     "@vitejs/plugin-react": "catalog:",
     "@vitejs/plugin-rsc": "catalog:",
     "@zgeoff/bun-test-extended": "catalog:",
