--- projects/lib-client-test-utils/src/test-setup/index.ts
@@ -1,1 +1,0 @@
-export { registerMSWLifecycle } from './register-msw-lifecycle';
